[MIRROR] Removes the accounting console board from the sec lathe and moves it to the service lathe [MDB IGNORE] (#18582)
* Removes the accounting console board from the sec lathe and moves it to the service lathe (#72276)

## About The Pull Request
Removes the design for the accounting console circuit board. There is
now a replacement circuit board in the HoP locker.

This PR is mainly meant to address a downstream issue, but I figure I
should make the PR up here first.
## Why It's Good For The Game
It's the HoPs job to monitor and handle the accounting console, not
security's. This will encourage security to co-operate and talk with the
HoP if they need to monitor purchases.
## Changelog
:cl:
balance: Moved the account console board to the service lathe
/:cl:

Co-authored-by: Fikou <23585223+Fikou@ users.noreply.github.com>

* Removes the accounting console board from the sec lathe and moves it to the service lathe

Co-authored-by: Pinta <68373373+softcerv@users.noreply.github.com>
Co-authored-by: Fikou <23585223+Fikou@ users.noreply.github.com>