[MIRROR] Organizes the lists in `admin_verbs.dm` because I got sick and tired of looking at them, and one in particular [MDB IGNORE] (#18307)
* Organizes the lists in `admin_verbs.dm` because I got sick and tired of looking at them, and one in particular (#72151)

Title for the most part.

I was told, many, MANY, moons go, as a downstream host, that one of
these lists would rarely ever get touched.
What a fat lie that was.
The more I had to stare at this list the more infuriated I got with it.

Now, I just went ahead and organized most of the lists in this file, as
well as fixing comments to fit the format of `/* hey bozo */`. Where
applicable, things were organized into `/datums/admins/` and
`/client/proc` as well.

:cl: Jolly
code: The lists in admin_verbs.dm was organized to be made better. This
shouldn't affect the panels admins use at all, since they're already
alphabetized.
/:cl:

* Literally already broke the sorting rules within 2 seconds

Co-authored-by: Jolly <70232195+Jolly-66@users.noreply.github.com>
Co-authored-by: Funce <funce.973@gmail.com>