[MIRROR] Removes Thunderdome Showtime Button Off Window [MDB IGNORE] (#18653)
* Removes Thunderdome Showtime Button Off Window (#72451)

## About The Pull Request

![image](https://user-images.githubusercontent.com/34697715/210458954-08e73e91-6d77-4dd4-b6fa-ef9fcfae9c29.png)

This wasn't even in the template version of the thunderdome, so any time
an admin would reload the thunderdome, it wouldn't show up again. It's
just silly to have buttons on windows regardless of context. (if you
actually used this, just toggle `Admin AI Interaction` and hit the
button yourself or something).
## Why It's Good For The Game

It doesn't look good on a window, I promise you.
## Changelog
:cl:
fix: There is no longer a button on a window at the CentCom Thunderdome.
/:cl:

* Removes Thunderdome Showtime Button Off Window

Co-authored-by: san7890 <the@san7890.com>