[MODULAR] Fixes self-referencing recipe 'battered sausage' (#18758)
* Fixes self-referencing recipe

No longer get stuck in crafting loops

* Changed crafting recipe again to make more sense, & fun extras

Changed recipe so that you
1) craft raw battered sausage first
2) grill raw sausage on griddle

from before
1) craft normal sausage first
2) grill sausage on griddle
3) craft beer battered sausage from normal cooked sausage

Added sprite for cooked version of beer battered sausages.

Added some nutriment to the cooked version to represent the added nutrition that battering the sausage brings over regular sausages.

Raw version has the uncooked batter in it.

Made beerbatter very slightly alchoholic, since it's made with beer. It says the scottish drink it so...

* Added 2 units of oil

For the added nutrition

* formatting

* Reformatted food_reagents lists to make them consistent with the rest of the code

Also clarified the description for battered sausages

* Further reformats the food_reagents lists in modular to match the rest of the code