[MIRROR] Adds the Chin-Length Bob Cut hairstyle! [MDB IGNORE] (#18828)
Adds the Chin-Length Bob Cut hairstyle! (#72788)

## About The Pull Request
Adds a gender neutral (with an albeit feminine touch) haircut! It
circles around the face, with it ending at more or less the same level
as the chin of a human!

![image](https://user-images.githubusercontent.com/77556824/213096317-cf506af9-9d36-4166-b61e-181a7c988296.png)


![image](https://user-images.githubusercontent.com/77556824/213096441-31ff7f6b-8e88-4613-a911-986f751a01b0.png)
## Why It's Good For The Game
More character flavor, and it was a haircut I've used two entire years
without ever formally adding it, and I've had the blessings of @Twaticus
to do so!
## Changelog
:cl: PositiveEntropy
add: Adds a new short length haircut: the Chin-Length Bob Cut!
/:cl:

Co-authored-by: Imaginos16 <77556824+Imaginos16@users.noreply.github.com>