[MIRROR] Life and Mana bars for the TGC battle arena. [MDB IGNORE] (#18934)
Life and Mana bars for the TGC battle arena. (#72838)

## About The Pull Request

Adds a life and mana bar for each player in the TGC holodeck arena,
practically removing the need for players to keep track of these stats
using items outside of the holodeck. The mana bar features a next turn
button which will automatically increase the number of slots by 1 and
refill or slots. You can inspect the panels to see these values
numerically rather than visually.


![image](https://user-images.githubusercontent.com/40036527/213697852-9743a309-6cf1-40d0-92d7-c3be3f1884e5.png)
## Why It's Good For The Game

I've set out on a goal on making the TGC battle arena a complete
visualizer for the TGC game, where all you need to bring is a deck of
cards and a friend. Adding this eliminates the need for pen and paper or
an external notepad program for tracking life and mana, and replaces it
with a far more visually interesting option that is also significantly
easier for both players to see and keep track of.
## Changelog
:cl:
add: The TGC arena on the holodeck now features display panels to track
each player's life shards and mana.
qol: Slightly increased the amount of space on the TGC holodeck arena's
tables by removing the windows on either side of the player.
/:cl:

Co-authored-by: NamelessFairy <40036527+NamelessFairy@users.noreply.github.com>