[MIRROR] Fixes the ORM I/O lights [MDB IGNORE] (#19012)
Fixes the ORM I/O lights (#72936)

## About The Pull Request

The ORM's input lights have turned invisible, due to item state names
that have not been updated. This PR fixes that. It also autodocs the
various procs and vars.
 

## Why It's Good For The Game
Fixes #72923 It is good to see which direction you have to place the
ores.

## Changelog
 

:cl: 
fix: The ORM's input and output lights are once again visible
/:cl:

Co-authored-by: Profakos <profakos@gmail.com>